commit #13
We cannot be sure that the article scraped will always be about the company,
Sometimes, only a few sentences will be about the company of our wish.
So I tried to introduce a mechanism to add the company to the context and extract summary based on that information.
It worked on paper. But due to resource limitation. System is running into BSOD errors.

Adding generate_summary v2
But need to intriduce a much less resource inducing function